fix(automation): move template picker query into business and scope reflink writes
Move the template picker's selectable-resource dispatch out of the builder
query layer into `templateService.listSelectableResources`, so the query file
is a thin adapter and the repository is no longer reached from `apps/`.

Scope `reflinkService` update/deleteMany/listOptions by `type = "refLink"` to
match what `create` stamps, so an entry-point-link row sharing the table can
never be updated or deleted through the reflink surface.

Stop swallowing every error on the flow pages: only a service-thrown
`notFoundException` becomes `notFound()`, and a DB failure propagates as a
real 500 instead of a misleading 404.

Drop the now-unused `broadcastRepository.findIdIfActive` and
`findContactFilter` along with their tests and service mocks.
