+++ apps/builder/src/app/(no-sidebar)/space/[workspaceId]/flows/[id]/analytics/page.tsx
@@ -10,6 +10,7 @@ import { FlowStoreProvider } from "@/features/flows/provider/flow-store-context"
 import { FlowTemplateStoreProvider } from "@/features/flows/react-flow/stores/flow-template-store-provider"
 import { withWorkspaceIdAndIdSchema } from "@/features/workspaces/schema/resource"
 import { requireWorkspacePermission } from "@/lib/auth/require-workspace-permission"
+import { isNotFoundException } from "@/lib/errors/validation-exception"
 
 type FlowAnalyticsPageProps = {
   params: Promise<{ workspaceId: string; id: string }>
@@ -31,8 +32,11 @@ export default async function FlowAnalyticsPage({
       id: data.id,
       workspaceId: data.workspaceId,
     })
-  } catch {
-    return notFound()
+  } catch (error) {
+    if (isNotFoundException(error)) {
+      return notFound()
+    }
+    throw error
   }
 
   const draftFlowVersion = flow.flowVersions?.find((v) => v.isDraft)
