+++ packages/business/__tests__/broadcast-service-resend.test.ts
@@ -78,7 +78,6 @@ vi.mock("@chatbotx.io/database/repositories", () => ({
   broadcastRepository: {
     listWithRelations: vi.fn(),
     count: vi.fn(),
-    findIdIfActive: vi.fn(),
     listAudience: vi.fn(),
     countAudience: vi.fn(),
     findByIdOrName: vi.fn(),
