+++ packages/business/__tests__/broadcast-service-update.test.ts
@@ -64,7 +64,6 @@ vi.mock("@chatbotx.io/database/repositories", () => ({
   broadcastRepository: {
     listWithRelations: vi.fn(),
     count: vi.fn(),
-    findIdIfActive: vi.fn(),
     listAudience: vi.fn(),
     countAudience: vi.fn(),
     findByIdOrName: vi.fn(),
